Rearrange folder structure into packages to move towards application-style architecture
