+++ backend/src/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Aug  3 08:48:30 2021
+
+@author: Pooja3894
+"""
+__version__ = "0.1"
