+++ backend/src/static_boundary_conditions/lidar/dem_metadata_in_db.py
