+++ backend/src/static_boundary_conditions/vector/tables.py
